\name{doesparql-package}
\alias{doesparql-package}
\alias{doesparql}
\docType{package}
\title{SPARQL Evaluations using Design of Experiments}
\description{The R code needed to run the experiments in the study titled "Introducing Statistical Design of Experiments to SPARQL Endpoint Evaluation" by Kjernsmo & Tyssedal, to reproduce the study. It also contains some utility functions, as well as code to do the actual timings.}
\details{
\tabular{ll}{
Package: \tab doesparql\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-27\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Kjetil Kjernsmo

Maintainer: Kjetil Kjernsmo <kjekje@ifi.uio.no>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ DoE }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
