\name{compose}
\alias{compose}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Help composing experiment.
}
\description{
This function just extracts the right part of the files. It is not of
interest to users.
}
\usage{
compose(factor, type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{factor}{
The present factor.
}
  \item{type}{
The type of the experimental factor.
}
}
\value{
Returns the correct part of the file if it is of the right type, NA otherwise.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Kjetil Kjernsmo
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

